Added '3D' mode to DemoCamera (enabled via property browser).
Added '3D' mode to DemoCamera (enabled via property browser).

I'm not entirely confident in the mathematics, but when enabled, the DemoCamera
will now project the corners of a cube as dots. They surround the 'stage'
positions and rotate around the 'upwards' axis much like the sample in the
OpenSPIM project does. It's still somewhat unstable, however; I've been
encountering various types of CTDs while it's enabled (still tracking them), so
play with it at your own risk. :) I've restored the original behavior of
DemoCamera, in favor of this approach.

Further work on DemoCamera's 3D mode.

In an effort to track down some of the internal bugs still plaguing the 3D
mode, I reorganized some of the code. The composite projection matrix seems to
be off; I added numbers to the corners of each dot to help track where they're
showing up.

Fixes for DemoCamera's 3D mode.

Apparently I'm composing my matrices wrong, because I've replaced that section
of code with the individual matrices applied in turn to the vectors, and they
seem to project correctly now.

DemoCamera 3D mode: Remembered to convert FOV to radians

Also reinstated dot scaling with depth, so now it's clear whether a particle is
in the background or foreground and the like. It now provides an excellent
environment in which to test acquisition.

Successfully composed the 3D-mode matrices.

Finally figured out the right order to  do it in. MATLAB helped a lot; call me
lazy, but I'd rather not multiply four 4x4 matrices on paper.

Added support for extra pixel types in the DemoCamera 3D mode.

3D mode now supports 16-bit, 32-bit float, and 32-bit ARGB pixel modes. No
64-bit RGB. I'm trying to mirror the capabilities of the original DemoCamera
implementation (2D), which supports all of the above modes.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>