Corrected release configuration for OpenSPIM DALs
The 2010 DAL project files have debug information enabled in release mode.
This presents a problem as we aren't allowed to redistribute the debug
runtime libraries. I've disabled the relevant setting for the OpenSPIM
DALs !ONLY!. Any device adapters we bring into the fold will need to be
corrected as well.

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>